Make Pacto::Contract a module, included by concrete format-specific Contracts
