+++ lib/pacto/formats/legacy/request_clause.rb
@@ -1,0 +1,9 @@
+# -*- encoding : utf-8 -*-
+module Pacto
+  module Formats
+    module Legacy
+      class RequestClause < Pacto::RequestClause
+      end
+    end
+  end
+end
