+++ lib/pacto/formats/legacy/response_clause.rb
@@ -1,0 +1,8 @@
+module Pacto
+  module Formats
+    module Legacy
+      class ResponseClause < Pacto::ResponseClause
+      end
+    end
+  end
+end
